% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{compute_marginal_likelihood}
\alias{compute_marginal_likelihood}
\title{Compute Marginal Likelihood}
\usage{
compute_marginal_likelihood(X, kernel_param, class_labels)
}
\arguments{
\item{X}{(matrix) : The design matrix of (S/D) EC curves}

\item{kernel_param}{(float) : Bandwidth parameter for the Gaussian Kernel}

\item{class_labels}{(vector) : vector of +1,-1 of class labels.}
}
\value{
log_Z (float): the EP log likelihood given the parameters.
}
\description{
\code{compute_marginal_likelihood} Computes Marginal likeilhood for the GPC Model, when using Expectation Prop.
Used for as model evidence in order to compute optimal parameters/ hyperparameters for the model. Implemented so far
only for the GaussKernel kernel function. Might be expanded in the future.
}
