% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{compute_rbf_model}
\alias{compute_rbf_model}
\title{Computes the RBF Kernel}
\usage{
compute_rbf_model(data, eta = 5, func)
}
\arguments{
\item{eta}{(float) The varying shape parameter.}

\item{x}{(vector) The first vector}

\item{y}{(vector) The second vector}
}
\value{
rbf_final (list) : The rbf model values: rbf weights, input data, and interpolation function
}
\description{
Computes the inverse quadratic distance between two vectors
}
