% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{create_data_normal_fixed}
\alias{create_data_normal_fixed}
\title{Generate (S/D) EC curves for a collection of interpolations on the plane}
\usage{
create_data_normal_fixed(
  num_sim = 25,
  dir,
  curve_length = 10,
  shared_points = 5,
  causal_points = 5,
  grid_size = 25,
  func = rbf_gauss,
  eta = 5,
  ball = TRUE,
  ball_radius = 1,
  ec_type = "ECT"
)
}
\arguments{
\item{num_sim}{(int) : The number of replicates of data.}

\item{dir}{(matrix) : The matrix of directions we compute the EC curves over.}

\item{curve_length}{(int) : Number of sub-level sets in each EC computation.}

\item{shared_points}{(int) : The number of shared points across both classes of shapes.}

\item{causal_points}{(int) : The number of class specific causal points in each class.}

\item{grid_size}{(int) : The fine-ness/granularity of the interpolated shapes.}

\item{func}{(function) : The function used to compute the interpolations.}

\item{eta}{(float) : The kernel shape parameter.}

\item{ball}{(boolean) : Denotes whether or not to compute the EC curves over a ball for uniform measurements}

\item{ball_radius}{(float) : The radius of the bounding ball used if we compute the balled EC curve.}

\item{ec_type}{(string) : The type of EC we are computing. We currently support ECT, DECT and SECT.}
}
\value{
data_list (list) : List of procedure specific data : The associated (S/D) EC matrix, coordinates of causal & shared points,
and the vertex coordinates of all the shapes.
}
\description{
Given a set of input parameters, \code{create_data_normal_fixed} generates replicates of interpolations on the plane.
These interpolations are then converted to simplicial complexes, and have the (S/D) EC curve computed over them.
}
