% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction_Functions.R
\name{critical_points_feature_selection}
\alias{critical_points_feature_selection}
\title{The Reconstruction Procedure, given the indices deemed important.}
\usage{
critical_points_feature_selection(critical_points, indices, direction)
}
\arguments{
\item{critical_points}{A list. Each entry are the critical points for each direction in matrix form, with indices in the first column.}

\item{indices}{The output of the feature selection algorithm. The ec curves in each direction are concatenated into a long vector, and feature
selection algo selects indices that are important for each direction.}

\item{direction}{a matrix of directions; should be the ones that generated the input critical points as well.}
}
\value{
Returns the coordinates of the selected critical points.
}
\description{
\code{critical_points_feature_selection} returns the coordinates of the critical points that are chosen by the feature selection procedure.
}
\details{
The code first checks if the number of directions in the critical points input and directions input match up. If so, we pick out all the
critical points that have indices contained within the desired indices, the output of the feature selection algorithm. Essentially, each critical
point comes with an identifier unique to the direction and index on that direction's ec curve. If that identifier matches with the index chosen
by the feature selection algorithm, we select that critical point. We take a union of sorts: if a critical point has a projection onto a direction
that is chosen by feature selection, we select that critical point.

This procedure can be changed, but at least should be discussed; we could take the intersection for example, but there is no guarantee that
any critical point satisfies this property. Could revive the evidence threshold idea.
}
