% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{difference}
\alias{difference}
\title{Euclidean Distance}
\usage{
difference(x, y)
}
\arguments{
\item{x}{(vector) : The first vector}

\item{y}{(vector) : The second vector}
}
\value{
The Euclidean distance
}
\description{
Computes the Euclidean distance between two vectors
}
