% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{ExpectationPropagation}
\alias{ExpectationPropagation}
\title{Use Expectation Propogation to Approximate mean & covariance}
\usage{
ExpectationPropagation(K, class_labels)
}
\arguments{
\item{K}{(matrix): the covariance matrix for the GP model}

\item{class_labels}{(vector): +/- 1 values indicating the class labels of the data points}
}
\value{
params (list): list of the posterior mean and variances.
}
\description{
\code{ExpectationPropogation} Approximates the latent posterior with a Gaussian distributions; it does so by moment matching.
Pseudocode taken from Rasmussen and Williams, Chapter 3. This function outputs the mean and covariance of the
approximated posterior. To actually generate samples from the latent posterior, generate samples from a multivariate
normal with the parameters returned by this function.
}
