% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction_Functions.R
\name{find_critical_points_curve}
\alias{find_critical_points_curve}
\title{Returns the critical points projections given a direction and the simplicial complex.}
\usage{
find_critical_points_curve(directions, off, len)
}
\arguments{
\item{directions}{A list. Each entry is a separate direction (a unit vector), in matrix form. This is the output of
\code{generate_perturbed_directions}.}

\item{off}{A simplicial complex. Can be obtained by calling methods from SECT on an OFF file.}

\item{len}{Integer. Length of the ec curve.}
}
\value{
The output is a list: each entry i consists of the projections of critical points onto direction i, defined as
the index where the ec curve in direction i changes value. The coordinate of the projection (the distance from the origin
to the projected) is also recorded.
}
\description{
\code{find_critical_points_curve} Returns a list of the values at which the ec curve changes, for every direction specified.
}
\details{
The function is to compute the critical point projections for a cone of directions.
Can be used more generally so long as the input is a list. For each direction, the function computes the unsmoothed Euler Characteristic
Curve and computes the indices and corresponding coordinate for which the EC changes.
}
