% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction_Functions.R
\name{find_critical_points_multiple_directions}
\alias{find_critical_points_multiple_directions}
\title{Finds critical points for multiple directions.}
\usage{
find_critical_points_multiple_directions(dir, off, steps, phi)
}
\arguments{
\item{dir}{A matrix of directions, n x 3. Each row contains a direction for which to compute the critical points.}

\item{off}{a simplicial complex: a list containing vertices, faces, and edges. Obtained by calling methods from the SECT_Functions file.}

\item{steps}{Integer. Length of the ec curve.}

\item{phi}{Float. Degree of the angle between input direction and the original.}
}
\value{
A list. Each entry corresponds to a different direction, and the entry is an n x 4 matrix, where n is the number of critical points found.
 The last three columns are the x,y,z coordinates of the critical points. The first column is the index in the ec curve that corresponds to that
 critical point; this is adjusted so that it matches the direction that the critical point came from, giving it a unique identifier in terms of which
 direction and where on the ec curve that critical point corresponds to.
}
\description{
\code{find_critical_points_multiple_directions} computes the coordinates of critical points given a list of directions and the desired simplicial
complex. The number of steps denotes the number of sublevel sets to consider when computing the EC curve, and phi is the initial angle.
}
