% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{GaussKernel}
\alias{GaussKernel}
\title{Gaussian Kernel Implementation}
\usage{
GaussKernel(X, bandwidth)
}
\arguments{
\item{X}{(p x n matrix): the design matrix where columns are observations}

\item{bandwidth}{(float): free parameter for the Gaussian kernel}
}
\description{
\code{GaussKernel} computes the covariance matrix given by the Gaussian kernel below.
}
