% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{generate_data_sphere_simulation}
\alias{generate_data_sphere_simulation}
\title{Generate (S/D) EC curves for a collection of perturbed spheres}
\usage{
generate_data_sphere_simulation(
  nsim,
  curve_length,
  dir,
  noise_points = 5,
  causal_points = 5,
  ball = TRUE,
  ball_radius = 2,
  ec_type = "ECT",
  subdivision = 3,
  cusps,
  causal_regions_1 = c(1),
  causal_regions_2 = c(3),
  shared_regions = c(4),
  write = FALSE,
  workdir = "~/Documents/spheres/v100"
)
}
\arguments{
\item{curve_length}{(int) : Number of sub-level sets in each EC computation.}

\item{dir}{(matrix) : The matrix of directions we compute the EC curves over.}

\item{noise_points}{(int) : The number of shared points in each shared cusp.}

\item{causal_points}{(int) : The number of class specific causal points in each causal cusp.}

\item{ball}{(boolean) : Denotes whether or not to compute the EC curves over a ball for uniform measurements}

\item{ball_radius}{(float) : The radius of the bounding ball used if we compute the balled EC curve.}

\item{ec_type}{(string) : The type of EC we are computing. We currently support ECT, DECT and SECT.}

\item{subdivision}{(int) : The fineness of the mesh. We currently use subdivision = 3.}

\item{cusps}{(int) : The number of total cusps to be generated.}

\item{causal_regions_1}{(vector) : The index of cusps to be denoted as the central points for the causal cusps of class 1.}

\item{causal_regions_2}{(vector) : The index of cusps to be denoted as the central points for the causal cusps of class 2.}

\item{shared_regions}{(vector) : The index of cusps to be denoted as the central points for the shared cusps for both classes.}

\item{num_sim}{(int) : The number of replicates of data.}
}
\value{
data_list (list) : List of procedure specific data : The associated (S/D) EC matrix, indices of the shared & causal vertices,
and the metadata for locations of the cusps.
}
\description{
Given a set of input parameters, \code{create_data_sphere_simulation} generates replicates of perturbed spheres.
Cusps are initialized by generating equidistributed points on the sphere. The cusps are created by taking the $k$ nearest vertices of these central points,
and depending on if the cusps are causal or shared, the cusps are indented or made protruding out.
These perturbed spheres are divided into different classes using the indented cusps.
These spheres then have the (S/D) EC curve computed over to create the design matrix $X$.
}
