% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction_Functions.R
\name{generate_perturbed_directions}
\alias{generate_perturbed_directions}
\title{Generates 4 directions in a small cone around a specified direction.}
\usage{
generate_perturbed_directions(direction, phi = 0.001)
}
\arguments{
\item{direction}{A vector. Should ideally be normalized so norm is one.}

\item{phi}{A scalar. The angle parameter for the angle between the given direction and its adjacent directions.}
}
\value{
The output is a list of length 5, with direction is one entry of the list.
}
\description{
\code{generate_perturbed_directions} returns a list of the input direction and the four directions.
}
\details{
The function first computes two orthogonal vectors to the input direction, then perturbs the original direction
by rotating the original direction an angle of phi and -phi in the direction of the two orthogonal vectors, which
generates four additional directions in a cone around the original.
}
