% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction_Functions.R
\name{get_all_critical_points}
\alias{get_all_critical_points}
\title{Extracts coordinates of critical points from a list.}
\usage{
get_all_critical_points(critical_point_list)
}
\arguments{
\item{critical_point_list}{A list. Each entry is a matrix of critical points, the first column are the indices of the ec curve and the remaining
columns are the coordinates of the critical points.}
}
\value{
A matrix of solely the coordinates of the critical points.
}
\description{
\code{get_all_critical_points} is a simple wrapper for \code{find_critical_points_multiple_directions}; it takes the output of the latter function
and extracts the coordinates of the critical points for each direction.
}
