% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{gradSqExpKernel}
\alias{gradSqExpKernel}
\title{Gaussian Kernel Implementation}
\usage{
gradSqExpKernel(X, bandwidth)
}
\arguments{
\item{X}{(p x n matrix): the design matrix where columns are observations}

\item{bandwidth}{(float): free parameter for the Gaussian kernel}
}
\description{
\code{gradUnstandardizedGaussKernel} computes the gradient covariance matrix given by the Gaussian kernel below, without
setting the diagonal to zero.
}
