% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{inv_quad}
\alias{inv_quad}
\title{Inverse Quadratic Kernel}
\usage{
inv_quad(x, y, eta = 5)
}
\arguments{
\item{x}{(vector) The first vector}

\item{y}{(vector) The second vector}

\item{eta}{(float) The varying shape parameter.}
}
\value{
The inverse quadratic distance
}
\description{
Computes the inverse quadratic distance between two vectors
}
