% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{logistic_log_likelihood}
\alias{logistic_log_likelihood}
\title{Probit Log Likelihood}
\usage{
logistic_log_likelihood(latent_variables, class_labels)
}
\description{
When the link function is logistic
}
