% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{matrix_to_simplicial_complex}
\alias{matrix_to_simplicial_complex}
\title{Convert 2d function to simplicial complex}
\usage{
matrix_to_simplicial_complex(matrix, grid_length)
}
\arguments{
\item{matrix}{(nx3 matrix) The matrix we want to convert into a simplicial complex}

\item{grid_length}{(float): The length of the grid for which the shape is defined on.}
}
\value{
complex (list) : List of metadata about our mesh; the vertex coordinates, edge positions, and face positions.
}
\description{
Using Rvcg, we turn a matrix into a triangular simplicial complex.
}
