% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{optimize_params_EP}
\alias{optimize_params_EP}
\title{Pick best kernel parameters}
\usage{
optimize_params_EP(X, class_labels, initial_params)
}
\arguments{
\item{X}{(matrix) : The design matrix of (S/D) EC curves}

\item{class_labels}{(vector) : vector of +1,-1 of class labels.}

\item{kernel_param}{(float) : Bandwidth parameter for the Gaussian Kernel}
}
\value{
Rcgmin output
}
\description{
\code{optimize_params_EP} Uses nonlinear conjugate gradient
optimizer to pick out the best value of the kernel params
}
