% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction_Functions.R
\name{plot_complex_with_points}
\alias{plot_complex_with_points}
\title{\code{plot_complex_with_points} Plot the 3d shape, and the critical points (convex hull or points themselves).}
\usage{
plot_complex_with_points(points, vertices, n = 100, main_color = "white",
  crit_color = "blue", alpha_one = 0.5, alpha_two = 0.3, off_file,
  axes = FALSE, labels = FALSE, plot_points = FALSE, hull = TRUE,
  point_size = 10)
}
\arguments{
\item{points}{A (nx3 matrix). Each entry are the critical points in matrix form. These points are what we create the convex hull around.}

\item{n}{(integer) A parameter to choose, how big the convex hull is (how many of the n closest vertices), if chosen.}

\item{main_color}{(string) Color of the shape.}

\item{crit_color}{(string) Color of the convex hull/shape.}

\item{alpha_two}{(float (0-1)): The intensity of the color of the shape.}

\item{off_file}{(mesh3d): A mesh3d file, read in from Rvcg. This is the shape that's plotted.}

\item{axes}{(TRUE/FALSE): A boolean to determine if the axes will be shown.}

\item{labels}{(TRUE/FALSE): A boolean to determine if the axes labels will be shown.}

\item{plot_points}{(TRUE/FALSE): A boolean to see if the user wants to plot the critical points themselves, in addition to the hull.}

\item{hull}{(TRUE/FALSE): A boolean to determine if the user wants to display the convex hull.}

\item{point_size}{(int): If the critical points are plotted, this specifies the point size.}

\item{alpha_two}{(float (0-1)): The intensity of the color of the convex hull (if chosen)}
}
\value{
Plots the 3d shape,and the plotting of the critical points/or the convex hull of the critical points.
}
\description{
The code plots the 3d shape, and the plotting of the critical points. There are a few plotting parameters, determining the shading and transparency of the shapes.
The critical points provided are then plotted too. They can be plotted either as points, or as a convex hull. These are given as boolean conditions.
}
