% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction_Functions.R
\name{plot_points}
\alias{plot_points}
\title{\code{plot_points} Plot the critical points .}
\usage{
plot_points(points, point_size = 5, point_color = "red", alpha = 1)
}
\arguments{
\item{points}{A (nx3 matrix). Each entry are the critical points in matrix form. These points are what we create the convex hull around.}

\item{point_size}{(int): Specifies the point size of the critical points.}

\item{point_color}{(string) Color of the shapes.}

\item{alpha}{(float (0-1)): The intensity of the color of the points.}
}
\value{
Plots the Critical points.
}
\description{
The code plots the critical points.
}
