% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RATE.R
\name{RATEv2}
\alias{RATEv2}
\title{X' is the nxp design matrix (e.g. genotypes) where n is the number of samples and p is the number of dimensions. This is the original input data
f.draws' is the Bxn matrix of the nonparametric model estimates (i.e. f.hat) with B being the number of sampled (posterior) draws;
nullify' is an optional vector specifying a given predictor variable effect that user wishes to "remove" from the data set. An example of this corresponds to Figures 2(b)-(d) in Crawford et al. (2018);
snp.nms' is an optional vector specifying the names of the predictor variables;
cores' is a parameter detailing the number of cores to parallelize over. If too many are assigned, RATE will set this variable to maximum number of cores that are available on the operating machine.}
\usage{
RATEv2(X, f.draws = NULL, pre.specify = FALSE, beta.draws = NULL,
  prop.var = 1, rank.r = min(nrow(X), ncol(X)), nullify = NULL,
  snp.nms = NULL, cores = 1)
}
\description{
X' is the nxp design matrix (e.g. genotypes) where n is the number of samples and p is the number of dimensions. This is the original input data
f.draws' is the Bxn matrix of the nonparametric model estimates (i.e. f.hat) with B being the number of sampled (posterior) draws;
nullify' is an optional vector specifying a given predictor variable effect that user wishes to "remove" from the data set. An example of this corresponds to Figures 2(b)-(d) in Crawford et al. (2018);
snp.nms' is an optional vector specifying the names of the predictor variables;
cores' is a parameter detailing the number of cores to parallelize over. If too many are assigned, RATE will set this variable to maximum number of cores that are available on the operating machine.
}
