% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{rbf_gauss}
\alias{rbf_gauss}
\title{RBF Kernel}
\usage{
rbf_gauss(x, y, eta = 5)
}
\arguments{
\item{x}{(vector) The first vector}

\item{y}{(vector) The second vector}

\item{eta}{(int) The varying weight parameter (not used here)}
}
\value{
The squared euclidean distance
}
\description{
Computes the squared euclidean distance between two vectors
}
