% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_directions.R
\name{rodriq}
\alias{rodriq}
\title{Compute the directions about a cone using the rodrigues angle formula.
 @export}
\usage{
rodriq(z, r, j)
}
\arguments{
\item{z}{(vector of length 3) The direction around which we want to rotate.}

\item{r}{(float): The radius of the cone we want. Controls the cone size.}

\item{j}{(int): The number of directions in the cone.}
}
\value{
dir (jx3 matrix): of directions in a cone of radius r about z.
}
\description{
Compute the directions about a cone using the rodrigues angle formula.
 @export
}
